backend/requirements.txt
fastapi
uvicorn
requests
passlib
pyjwt
python-multipart
beautifulsoup4
